Added feature: download all episodes from a list page (season)

+++ watch-dl.py
@@ -6,6 +6,7 @@ import re
 import urllib2
 import sys
 import os
+import os.path
 
 def info_extractor(url):
     _VALID_URL = r'(?:https://)?(?:www\.)?watchcartoononline\.io/([^/]+)'
@@ -15,7 +16,7 @@ def info_extractor(url):
         user_agent = 'Mozilla/5.0 (Windows NT 5.1; rv:10.0.1) Gecko/20100101 Firefox/10.0.1'
         headers = { 'User-Agent' : user_agent }
         
-        print "[watchcartoononline-dl]  Opening webpage"
+        print "[watchcartoononline-dl]  Downloading webpage"
         request = urllib2.Request(url,headers=headers)
         webpage = urllib2.urlopen(request).read()
     
@@ -24,7 +25,7 @@ def info_extractor(url):
         video_url = re.search('src="(.+?)"', video_url).group(1).replace(' ','%20')
         
         # "clicks" the "Click Here to Watch Free" button to so it can access the actual video file url
-        print "[watchcartoononline-dl]  Clicking stupid 'Watch Free' button"
+        #print "[watchcartoononline-dl]  Clicking stupid 'Watch Free' button"
         params = urllib.urlencode({'fuck_you':'','confirm':'Click Here to Watch Free!!'})
     
         print "[watchcartoononline-dl]  Getting video URL"
@@ -40,15 +41,64 @@ def info_extractor(url):
     else:
         print "ERROR: URL was invalid, please use a valid URL from www.watchcartoononline.com"
 
-def downloader(fileurl,file_name):
-    #opens the video file url
-    u = urllib2.urlopen(fileurl)
-    #writes new file with the filename provided
-    f = open(file_name, 'wb')
+def episodes_extractor(episode_list):
+    _VALID_URL = r'(?:https://)?(?:www\.)?watchcartoononline\.io/anime/([^/]+)'
+    #check if url is valid
+    if re.match(_VALID_URL, episode_list) is not None:
+    
+        #sets user_agent so watchcartoononline doesn't cause issues
+        user_agent = 'Mozilla/5.0 (Windows NT 5.1; rv:10.0.1) Gecko/20100101 Firefox/10.0.1'
+        headers = { 'User-Agent' : user_agent }
+        
+        print "[watchcartoononline-dl]  Downloading webpage"
+        request = urllib2.Request(episode_list, headers=headers)
+        webpage = urllib2.urlopen(request).read()
+
+        print "[watchcartoononline-dl]  Finding episode(s)"
+        
+        #remove the end of the html, to avoid matching episodes in the 'recenly added' bar
+        indexOfRecenly = webpage.find("Recenly")
+        truncated = ""
+        if indexOfRecenly != -1:
+            truncated = webpage[:indexOfRecenly]
+        else:
+            print "WARNING: couldn't find 'Recenly Added' section in page, maybe the site layout has changed?"
+            
+        #todo: improve this regex to work for more stuff
+        page_urls = re.findall(r'https://www.watchcartoononline.io/[a-zA-Z0-9-]+episode-[0-9]{1,4}[a-zA-Z0-9-]+', truncated)
+        #print list of URLs we are about to download
+        print "URLs found:"
+        for url in page_urls:
+            print url
+        
+        #run original script on each episode URL we found
+        for url in page_urls:
+            print "[watchcartoononline-dl]  Downloading "+url
+            doAnEpisode(url)
+    else:
+        print "ERROR: URL was invalid, please use a valid URL from www.watchcartoononline.com"
+
+def downloader(fileurl, file_name):
+    try:
+        #opens the video file url
+        u = urllib2.urlopen(fileurl)
+    except urllib2.HTTPError as he:
+        print "HTTPError! code:"+str(he.code)
+        return
+        
     #gets metadata
     meta = u.info()
     file_size = int(meta.getheaders("Content-Length")[0])
     file_type = meta.getheaders("Content-Type")[0]
+    
+    #before downloading, check if file already exists and is the expected size 
+    if os.path.isfile(file_name) and os.path.getsize(file_name) == file_size:
+        print "[watchcartoononline-dl]  file already exists and is the correct size, skipping..."
+        return
+
+    #writes new file with the filename provided
+    f = open(file_name, 'wb')
+
     print "[watchcartoononline-dl]  Filetype: %s" %(file_type)
     print "[watchcartoononline-dl]  Destination: %s" %(file_name)
     file_size_dl = 0
@@ -98,17 +148,27 @@ def convertSize(n, format='%(value).1f %(symbol)s', symbols='customary'):
             return format % locals()
     return format % dict(symbol=symbols[0], value=n)
 
+
+def doAnEpisode(url):
+    #url = sys.argv[1]
+    final_url = info_extractor(url)
+    if final_url is None:
+        print "ERROR: unable to extract video url from "+url
+    else:
+        name = final_url.replace('%20',' ').split('/')[-1]
+        name = name[:name.find('?')] # remove trailing URL arguments
+        downloader(final_url, name)
+
 if __name__ == '__main__':
     if len(sys.argv[1:]) > 0:
         try:
             url = sys.argv[1]
-            final_url = info_extractor(url)
-            if final_url is None:
-                print "ERROR: Try again"
-            else:
-                name = final_url.replace('%20',' ').split('/')[-1]
-                downloader(final_url,name)
-        #throws error message when keyboard inturupted eg: ctrl+c
+            if "/anime/" in url: #argument looks like an episode-list page
+                print "[watchcartoononline-dl] looks like a list of episodes (season?), extracting episode page URLs..."
+                episodes_extractor(sys.argv[1])
+            else: #episode should be a video page
+                doAnEpisode(url)
+        #throws error message for keyboard interrupt eg: ctrl+c
         except KeyboardInterrupt:
             print "\nERROR: Interrupted by user"
             try:
@@ -119,3 +179,5 @@ if __name__ == '__main__':
         #Prints some info if there was no argument
         print "Usage: python watch-dl.py [URL...]" 
         print "ERROR: You must provide a valid URL from www.watchcartoononline.com"
+
+
